fix: lint (#26)
There are some linting errors when adding experimentation to a
boilerplate project:


https://github.com/adobe-rnd/aem-boilerplate-xwalk/actions/runs/9174200125/job/25224528750?pr=22

This is caused by some fixable errors in `index.js` but also by the
`ued.js`. This file is listed in `.eslintignore`. That works here but
not in a project as `.eslintignore` must be in the project root:
https://eslint.org/docs/latest/use/configure/ignore-deprecated#the-eslintignore-file.
Adding a `/* eslint-disable */` to the file helps.